requirements/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
flake8==7.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.4.3
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pangres==4.2.1
pendulum==3.1.0
propcache==0.3.1
psycopg2==2.9.10
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
xlrd==2.0.1
yarl==1.19.0
